Move buildTarget out of src
