+++ tsconfig.json
@@ -22,5 +22,5 @@
       "@pages/*": ["src/pages/*"]
     }
   },
-  "include": ["src", "utils", "vite.config.ts"]
+  "include": ["src", "utils", "vite.config.ts", "buildTarget.ts"]
 }
